Fixed #6 - added RBAC and ABAC authorization layer and updated test data

+++ src/services/role/role.controller.ts
@@ -1,11 +1,17 @@
 import { NextFunction, Request, Response, Router } from "express";
 import { getRepository, Repository } from "typeorm";
-import { Role } from "./role.entity";
+import AuthPermission from '../../interfaces/permission.interface';
 import Controller from '../../interfaces/controller.interface';
-import CreateRoleDto from "./role.dto";
+import RequestWithUser from "../../interfaces/request.interface";
 import RecordNotFoundException from '../../exceptions/RecordNotFoundException';
+import RecordsNotFoundException from '../../exceptions/RecordsNotFoundException';
+import UserNotAuthorizedException from "../../exceptions/UserNotAuthorizedException";
 import authenticationMiddleware from '../../middleware/authentication.middleware';
 import validationMiddleware from '../../middleware/validation.middleware';
+import { methodActions, getPermission } from "../../utils/authorization.helper";
+
+import { Role } from "./role.entity";
+import CreateRoleDto from "./role.dto";
 
 /**
  * Handles CRUD operations on Role data in database
@@ -13,6 +19,7 @@ import validationMiddleware from '../../middleware/validation.middleware';
 class RoleController implements Controller {
   public path: string = "/roles";
   public router: Router = Router();
+  private resource: string = "role";
   private roleRepository: Repository<Role> = getRepository(Role);
 
   constructor() {
@@ -29,35 +36,76 @@ class RoleController implements Controller {
       .delete(`${this.path}/:id`, this.remove)
   }
 
-  private all = async (request: Request, response: Response, next: NextFunction) => {
+  private all = async (request: RequestWithUser, response: Response, next: NextFunction) => {
     const roles = await this.roleRepository.find({ relations: ["permissions"] });
-    response.send(roles);
-  }
+    
+    const isOwnerOrMember: boolean = false;
+    const action: string = methodActions[request.method];
+    const permission: AuthPermission = await getPermission(request.user, isOwnerOrMember, action, this.resource);
 
-  private one = async (request: Request, response: Response, next: NextFunction) => {
-    const { id } = request.params;
-    const role = await this.roleRepository.findOne(id, { relations: ["permissions"] });
-    if (role) {
-      response.send(role);
+    if (permission.granted) {
+      if (!roles) {
+        next(new RecordsNotFoundException(this.resource));
+      } else {
+        response.send(permission.filter(roles));
+      }
     } else {
-      next(new RecordNotFoundException(id));
+      next(new UserNotAuthorizedException(request.user.id, action, this.resource));
     }
   }
 
-  private save = async (request: Request, response: Response, next: NextFunction) => {
-    const roleData: CreateRoleDto = request.body;
-    await this.roleRepository.save(roleData);
-    response.send(roleData);
+  private one = async (request: RequestWithUser, response: Response, next: NextFunction) => {
+    const { id } = request.params;
+    const role = await this.roleRepository.findOne(id, { relations: ["permissions"] });
+
+    const isOwnerOrMember: boolean = false;
+    const action: string = methodActions[request.method];
+    const permission: AuthPermission = await getPermission(request.user, isOwnerOrMember, action, this.resource);
+
+    if (permission.granted) {
+      if (!role) {
+        next(new RecordNotFoundException(id));
+      } else {
+        response.send(permission.filter(role));
+      }
+    } else {
+      next(new UserNotAuthorizedException(request.user.id, action, this.resource));
+    }
   }
 
-  private remove = async (request: Request, response: Response, next: NextFunction) => {
-    const { id } = request.params;
-    const roleToRemove = await this.roleRepository.findOne(id);
-    if (roleToRemove) {
-      await this.roleRepository.remove(roleToRemove);
-      response.send(200);
+  private save = async (request: RequestWithUser, response: Response, next: NextFunction) => {
+    const roleData: CreateRoleDto = request.body;
+
+    const isOwnerOrMember: boolean = false;
+    const action: string = methodActions[request.method];
+    const permission: AuthPermission = await getPermission(request.user, isOwnerOrMember, action, this.resource);
+
+    if (permission.granted) {
+      const filteredData: CreateRoleDto = permission.filter(roleData);
+      await this.roleRepository.save(filteredData);
+      response.send(filteredData);
     } else {
-      next(new RecordNotFoundException(id));
+      next(new UserNotAuthorizedException(request.user.id, action, this.resource));
+    }
+  }
+
+  private remove = async (request: RequestWithUser, response: Response, next: NextFunction) => {
+    const { id } = request.params;    
+    const roleToRemove = await this.roleRepository.findOne(id);
+
+    const isOwnerOrMember: boolean = request.user.id === id;
+    const action: string = methodActions[request.method];
+    const permission: AuthPermission = await getPermission(request.user, isOwnerOrMember, action, this.resource);
+
+    if (permission.granted) {
+      if (roleToRemove) {
+        await this.roleRepository.remove(roleToRemove);
+        response.send(200);
+      } else {
+        next(new RecordNotFoundException(id));
+      }
+    } else {
+      next(new UserNotAuthorizedException(request.user.id, action, this.resource));
     }
   }
